+++ dune/localfunctions/lagrange/pqktracefactory.hh
@@ -132,18 +132,14 @@ namespace Dune
     /** \brief Copy constructor */
     PQkTraceLocalFiniteElementCache(const PQkTraceLocalFiniteElementCache& other)
     {
-      typename FEMap::iterator it = other.cache_.begin();
-      typename FEMap::iterator end = other.cache_.end();
-      for(; it!=end; ++it)
-        cache_[it->first] = (it->second)->clone();
+      for(const auto& entry : other.cache_)
+        cache_[entry.first] = (entry.second)->clone();
     }
 
     ~PQkTraceLocalFiniteElementCache()
     {
-      typename FEMap::iterator it = cache_.begin();
-      typename FEMap::iterator end = cache_.end();
-      for(; it!=end; ++it)
-        delete it->second;
+      for(auto&& entry : cache_)
+        delete entry.second;
     }
 
     //! Get local finite element for given GeometryType